+++ delta-coverage-core/src/main/kotlin/io/github/surpsg/deltacoverage/report/console/RawCoverageDataProvider.kt
@@ -1,0 +1,6 @@
+package io.github.surpsg.deltacoverage.report.console
+
+internal interface RawCoverageDataProvider {
+
+    fun obtainData(): List<RawCoverageData>
+}