+++ delta-coverage-core/src/main/kotlin/io/github/surpsg/deltacoverage/report/intellij/report/XmlReportBuilder.kt
@@ -6,14 +6,13 @@ import io.github.surpsg.deltacoverage.report.ReportBound
 import java.io.File
 
 internal class XmlReportBuilder(
-    reportsConfig: ReportsConfig,
-    reportBound: ReportBound,
-    reporter: Reporter,
-) : ReportBuilder(reporter, reportBound, reportsConfig) {
+    val reportBound: ReportBound,
+    private val reportsConfig: ReportsConfig,
+    private val reporter: Reporter,
+) : ReportBuilder {
 
     override fun buildReport() {
         val reportPath: File = ReportPathStrategy.Xml(reportsConfig).buildReportPath(reportBound)
         reporter.createXMLReport(reportPath)
     }
 }
-