+++ delta-coverage-core/src/main/kotlin/io/github/surpsg/deltacoverage/report/jacoco/analyzable/AnalyzableReport.kt
@@ -17,7 +17,6 @@ internal interface AnalyzableReport {
 internal fun analyzableReportFactory(
     reportContext: ReportContext
 ): Set<AnalyzableReport> {
-    reportContext.deltaCoverageConfig.coverageRulesConfig
     return reportFactory(reportContext)
         .map { reportMode ->
             when (reportMode) {