+++ delta-coverage-core/src/main/kotlin/io/github/surpsg/deltacoverage/report/jacoco/analyzable/FullCoverageAnalyzableReport.kt
@@ -2,8 +2,10 @@ package io.github.surpsg.deltacoverage.report.jacoco.analyzable
 
 import io.github.surpsg.deltacoverage.report.ConsoleHtmlReportLinkRenderer
 import io.github.surpsg.deltacoverage.report.FullReport
+import io.github.surpsg.deltacoverage.report.JacocoReport
 import io.github.surpsg.deltacoverage.report.ReportBound
 import io.github.surpsg.deltacoverage.report.ReportType
+import io.github.surpsg.deltacoverage.report.jacoco.csv.ConsoleCoverageReportOutputStream
 import org.jacoco.core.analysis.Analyzer
 import org.jacoco.core.analysis.ICoverageVisitor
 import org.jacoco.core.data.ExecutionDataStore
@@ -16,22 +18,37 @@ import org.jacoco.report.xml.XMLFormatter
 import java.io.File
 import java.io.FileOutputStream
 
+
 internal open class FullCoverageAnalyzableReport(
     private val report: FullReport
 ) : AnalyzableReport {
 
     override fun buildVisitor(): IReportVisitor {
-        return report.jacocoReports.map {
-            val reportFile: File = report.resolveReportAbsolutePath(it)
-            when (it.reportType) {
-                ReportType.HTML -> {
-                    ConsoleHtmlReportLinkRenderer.render(reportBound, reportFile)
-                    FileMultiReportOutput(reportFile).let(HTMLFormatter()::createVisitor)
-                }
-                ReportType.XML -> reportFile.createFileOutputStream().let(XMLFormatter()::createVisitor)
-                ReportType.CSV -> reportFile.createFileOutputStream().let(CSVFormatter()::createVisitor)
+        return report.jacocoReports
+            .mapNotNull { buildReportVisitor(it) }
+            .let(::MultiReportVisitor)
+    }
+
+    private fun buildReportVisitor(jacocoReport: JacocoReport): IReportVisitor? {
+        val reportFile: File = report.resolveReportAbsolutePath(jacocoReport)
+        return when (jacocoReport.reportType) {
+            ReportType.XML -> reportFile.createFileOutputStream().let(XMLFormatter()::createVisitor)
+
+            ReportType.CSV -> reportFile.createFileOutputStream().let(CSVFormatter()::createVisitor)
+
+            ReportType.HTML -> {
+                ConsoleHtmlReportLinkRenderer.render(reportBound, reportFile)
+                FileMultiReportOutput(reportFile).let(HTMLFormatter()::createVisitor)
             }
-        }.let(::MultiReportVisitor)
+
+            ReportType.CONSOLE -> {
+                if (reportBound == ReportBound.FULL_REPORT) {
+                    null
+                } else {
+                    ConsoleCoverageReportOutputStream(System.out).let(CSVFormatter()::createVisitor)
+                }
+            }
+        }
     }
 
     private fun File.createFileOutputStream(): FileOutputStream {
@@ -47,4 +64,5 @@ internal open class FullCoverageAnalyzableReport(
     }
 
     open val reportBound: ReportBound = ReportBound.FULL_REPORT
+
 }