+++ delta-coverage-core/src/test/kotlin/io/github/surpsg/deltacoverage/report/intellij/report/CoverageReportFactoryTest.kt
@@ -10,11 +10,15 @@ import io.github.surpsg.deltacoverage.report.intellij.coverage.NamedReportLoadSt
 import io.kotest.assertions.assertSoftly
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.equals.Equality
-import io.kotest.equals.byReflectionUsingFields
+import io.kotest.equals.EqualityResult
+import io.kotest.equals.ReflectionUsingFieldsEquality
+import io.kotest.equals.SimpleEqualityResult
+import io.kotest.equals.SimpleEqualityResultDetail
 import io.kotest.matchers.collections.shouldContain
 import io.kotest.matchers.collections.shouldHaveSize
 import io.kotest.matchers.sequences.shouldBeEmpty
 import org.junit.jupiter.api.Test
+import kotlin.reflect.KProperty
 
 class CoverageReportFactoryTest {
 
@@ -44,6 +48,7 @@ class CoverageReportFactoryTest {
             html = ReportConfig { enabled = true }
             xml = ReportConfig { enabled = true }
             csv = ReportConfig { enabled = true }
+            console = ReportConfig { enabled = true }
         }
 
         // WHEN
@@ -76,6 +81,7 @@ class CoverageReportFactoryTest {
         val config = ReportsConfig {
             html = ReportConfig { enabled = true }
             xml = ReportConfig { enabled = true }
+            console = ReportConfig { enabled = true }
         }
         val reportLoadStrategy = anyReportLoadStrategy()
 
@@ -87,16 +93,16 @@ class CoverageReportFactoryTest {
 
         // THEN
         assertSoftly(actualBuilders.toList()) {
-            shouldHaveSize(2)
+            shouldHaveSize(3)
 
             shouldContain(
                 HtmlReportBuilder(
                     REPORT_NAME,
-                    config,
                     REPORT_BOUND,
+                    config,
                     Reporter(reportLoadStrategy.reportLoadStrategy)
                 ),
-                Equality.byReflectionUsingFields(
+                EqualByFields.fromFields(
                     HtmlReportBuilder::reportName,
                     HtmlReportBuilder::reportBound,
                 )
@@ -104,11 +110,19 @@ class CoverageReportFactoryTest {
 
             shouldContain(
                 XmlReportBuilder(
+                    REPORT_BOUND,
                     config,
+                    Reporter(reportLoadStrategy.reportLoadStrategy)
+                ),
+                EqualByFields.fromFields(XmlReportBuilder::reportBound)
+            )
+
+            shouldContain(
+                ConsoleReportBuilder(
                     REPORT_BOUND,
                     Reporter(reportLoadStrategy.reportLoadStrategy)
                 ),
-                Equality.byReflectionUsingFields(HtmlReportBuilder::reportBound)
+                EqualByFields.fromFields(ConsoleReportBuilder::reportBound)
             )
         }
     }
@@ -123,6 +137,30 @@ class CoverageReportFactoryTest {
             Filters.EMPTY
         ),
     )
+
+    class EqualByFields<T : Any>(
+        private val fields: Array<out KProperty<*>>
+    ) : Equality<T> {
+
+        override fun name(): String = "Objects equal by instance and fields"
+
+        override fun verify(actual: T, expected: T): EqualityResult {
+            return if (actual::class.isInstance(expected)) {
+                ReflectionUsingFieldsEquality<T>(fields).verify(actual, expected)
+            } else {
+                SimpleEqualityResult(
+                    false,
+                    SimpleEqualityResultDetail {
+                        "Not the same type: actual=${actual::class}, expected=${expected::class}"
+                    }
+                )
+            }
+        }
+
+        companion object {
+            fun <T : Any> fromFields(vararg fields: KProperty<*>) = EqualByFields<T>(fields)
+        }
+    }
 
     private companion object {
         const val REPORT_NAME = "any-report-name"